오류 수정 중 실수 계산으로 바꾸기 전 커밋 NShape, Attribute, Symbol, Line, A4Paper 수정한 상태임.

+++ FlowChartEditorQt/FlowChart/Line.cpp
@@ -7,7 +7,7 @@
 
 #include "Line.h"
 
-Line::Line(Long x, Long y, Long width, Long height, QColor backGroundColor,
+Line::Line(float x, float y, float width, float height, QColor backGroundColor,
 	Qt::PenStyle borderLine, QColor borderColor, String contents)
 	: NShape(x, y, width, height, backGroundColor, borderLine, borderColor, contents) {
 }
