+++ FlowChartEditorQt/FlowChart/Symbol.cpp
@@ -7,7 +7,7 @@
 
 #include "Symbol.h"
 
-Symbol::Symbol(Long x, Long y, Long width, Long height, QColor backGroundColor,
+Symbol::Symbol(float x, float y, float width, float height, QColor backGroundColor,
 	Qt::PenStyle borderLine, QColor borderColor, String contents)
 	: NShape(x, y, width, height, backGroundColor, borderLine, borderColor, contents) {
